+++ src/main/scala/com/particeep/api/UserClient.scala
@@ -1,55 +1,12 @@
 package com.particeep.api
 
-import java.time.ZonedDateTime
-
 import com.particeep.api.core.{ResponseParser, WSClient}
 import play.api.libs.json._
 
 import scala.concurrent.{ExecutionContext, Future}
 import com.particeep.api.models._
-import com.particeep.api.models.enums.Gender.Gender
 import com.particeep.api.utils.LangUtils
-import com.particeep.api.core.Formatter
-
-case class User(
-  id:                       String                = "",
-  email:                    Option[String]        = None,
-  password:                 Option[String]        = None,
-  first_name:               Option[String]        = None,
-  last_name:                Option[String]        = None,
-  gender:                   Option[Gender]        = None,
-  avatar_url:               Option[String]        = None,
-  birthday:                 Option[ZonedDateTime] = None,
-  birth_place:              Option[String]        = None,
-  phone:                    Option[String]        = None,
-  nationality:              Option[String]        = None,
-  bio:                      Option[String]        = None,
-  sector:                   Option[String]        = None,
-  linkedin_url:             Option[String]        = None,
-  viadeo_url:               Option[String]        = None,
-  allow_mail_notifications: Option[Boolean]       = None,
-  has_been_claimed:         Option[Boolean]       = None,
-  address:                  Option[Address]       = None
-)
-
-case class UserSearchCriteria(
-  name:          Option[String] = None,
-  gender:        Option[String] = None,
-  birth_place:   Option[String] = None,
-  nationality:   Option[String] = None,
-  sector:        Option[String] = None,
-  tags:          Option[String] = None,
-  sort_by:       Option[String] = None,
-  order_by:      Option[String] = Some("asc"),
-  global_search: Option[String] = None,
-  limit:         Option[Int]    = Some(30),
-  offset:        Option[Int]    = Some(0)
-)
-
-object User {
-  implicit val date_format = Formatter.ZonedDateTimeWrites
-  implicit val format = Json.format[User]
-}
+import com.particeep.api.models.user.{User, UserCreation, UserEdition, UserSearch}
 
 trait UserCapability {
   self: WSClient =>
@@ -61,6 +18,8 @@ class UserClient(ws: WSClient) extends ResponseParser {
 
   private val endPoint: String = "/user"
   implicit val format = User.format
+  implicit val creation_format = UserCreation.format
+  implicit val edition_format = UserEdition.format
 
   def byId(id: String, timeout: Long = -1)(implicit exec: ExecutionContext): Future[Either[ErrorResult, User]] = {
     ws.url(s"$endPoint/$id", timeout).get().map(parse[User])
@@ -81,19 +40,19 @@ class UserClient(ws: WSClient) extends ResponseParser {
     ws.url(s"$endPoint/name/$name", timeout).get.map(parse[List[User]])
   }
 
-  def search(criteria: UserSearchCriteria, timeout: Long = -1)(implicit exec: ExecutionContext): Future[Either[ErrorResult, PaginatedSequence[User]]] = {
+  def search(criteria: UserSearch, timeout: Long = -1)(implicit exec: ExecutionContext): Future[Either[ErrorResult, PaginatedSequence[User]]] = {
     ws.url(s"$endPoint/search", timeout)
       .withQueryString(LangUtils.productToQueryString(criteria): _*)
       .get
       .map(parse[PaginatedSequence[User]])
   }
 
-  def create(user: User, timeout: Long = -1)(implicit exec: ExecutionContext): Future[Either[ErrorResult, User]] = {
-    ws.url(s"$endPoint", timeout).put(Json.toJson(user)).map(parse[User])
+  def create(user_creation: UserCreation, timeout: Long = -1)(implicit exec: ExecutionContext): Future[Either[ErrorResult, User]] = {
+    ws.url(s"$endPoint", timeout).put(Json.toJson(user_creation)).map(parse[User])
   }
 
-  def update(id: String, user: User, timeout: Long = -1)(implicit exec: ExecutionContext): Future[Either[ErrorResult, User]] = {
-    ws.url(s"$endPoint/$id", timeout).post(Json.toJson(user)).map(parse[User])
+  def update(id: String, user_edition: UserEdition, timeout: Long = -1)(implicit exec: ExecutionContext): Future[Either[ErrorResult, User]] = {
+    ws.url(s"$endPoint/$id", timeout).post(Json.toJson(user_edition)).map(parse[User])
   }
 
   def authenticate(email: String, password: String, timeout: Long = -1)(implicit exec: ExecutionContext): Future[Either[ErrorResult, User]] = {
